Shorten and fix (?) offset conversion to integer
This previous code, while it might have worked, seems to be an
oversight. It's easy to miss, but there is an elsif there, which means
the assignment happens always, not only when the condition returns a
truthy value.

I find a nice, short alternative for converting everything to an
integer, so this should do the trick. :)