[OSAP-1053] Forum stats about threads and posts
As a first step in simplifying the data model in the pinboard service,
the stats now don't make the distinction anymore between threads and
posts. While we still collect them separately, the service will expose
them combined to the outside.

Next step: use these combined values in the various admin dashboards and
in statistic mails.