Replace file service storage with S3
Store generated reports not on locale disk / xikolo-file service, but
use the S3 api to upload the file to cloud storage.

The ReportJob model has been extended by a download_url. For now
this will be used to store a presigned download url for the S3 object.
This might be refactored later, to be able to generated longer valid
URLs.

The S3 configuration is received via Xikolo.config. This include
connect information and the bucket to use. This services assumes
that the bucket has already been created and life cycle configuration
has been configured manually to automatic removed old reports.
The files should be keep at least seven days and afterwards be deleted.

The file_expiry_date is hardcoded to match the invalidity date of the
presigend-download url. The actual file expiration of the object
should be configured to be slightly longer, to the reports will be
removed from S3 a little bit after the URL got invalid.