Fix filter for long quiz submissions
Because we cannot measure precisely how long people actually spend on
quizzes, we use the time from first quiz access to time of submission as
a proxy. That is quite incorrect for selftests, though - as just opening
the quiz already creates the submission.

Therefore, we ignore all submissions that take longer than 3 hours. This
commit fixes that logic so that the threshold is at 3, not 30 hours. :)