Store retry options on RetryablePaginator
Because the constructor now accepts its own block with the request for
the first page (needed because that request needs to be retryable), the
`retryable_paginate` helper can not accept the pagination block also.

Instead, `each_item` has to be called explicitly (with a block) on the
returned paginator instance.