Add sentry error reporting
This commits adds and configures sentry error reporting. Existing error
attachments with mnemosyne have been extended to include an error report
to sentry. A new REVISION file is included and installed in the package
that includes the current revision that is packaged. The CI that builds
the package will override the REVISION file with the correct commit sha.

Sentry raven is configured to use a concurrent-ruby thread pool to
asynchronously send error reports.