Upgrade sidekiq and reconfigure redis
Switch to newest sidekiq major version. It is a straight forward
upgrade and on the long run we have to update at somepoint.

*Drop redis namespace (IS-725)*

Development and production heavily used `redis-namespace` to separate
data from different services and environments. The gem extends the
redis ruby interface to prepend a specific prefix for each key. This
is not possible for all operations and introduces performances
penalties. Its usage is disencourage by the sidekiq maintains since
many years. It is recommend to use different database or even
different redis installation for each purpose.

Our production installations have been updated to use a individual
Redis database per service. A additional Redis instance is used as
cache.

The local configuration files must be adjusted, too. The `namespace:`
keys should be removed and a different `db`-ID should be configured
per environment. Each service can use the database id that correspond
to its assigned port range (59xx for this service).
As development and production should use the same data on local installations
both use the same ID.

Updating sidekiq to 6.x and remove `redis-namespace` is bundled in
the `xikolo-sidekiq` 3.x release.

Furthermore, the redis configuration for integration is adjusted to
use only a database index.