Another fix for video statistics metric
The farthest watched metric causes issues for the other metrics when all
items of a courses are included in the query. The removal of the size
param fixed the other issues but caused wrong results of the farthest
watched metric. Therefore, we now request all metrics for each item of
a course separately. This should ensure to get correct results for each
metric.