Report metrics: Only add summary to error events
On production machines, we've discovered a few error messages in the
Telegraf logs. We traced this down to the "failing" events containing
complete exception stack traces. This is unnecessary and problematic. We
don't know the exact problem, but are assuming it's either because of
wrong formatting (unescaped newlines!) or size (too long for e.g. UDP).