Reports: Lock private tmp dir to prevent ageing
Adding a read lock to the temporary directory exempts it from systemd's
temp file ageing mechanism for as long as the file is opened.

Once the file handle is closed (even when the process is terminated
unexpectedly, e.g. because of our lovely out-of-memory errors), the
directory will again be considered for ageing, which avoids the need for
manual cleanup.

See https://systemd.io/TEMPORARY_DIRECTORIES/ for detailed explanation
and alternative options.

Refs #OWHO-268.