Remove deanonymization from unconfirmed user report
This was only applied to the User ID header, but not the data. It
doesn't make sense to provide pseudonymization for this report and we
actually never did so this is only an internal change. To help users
confirm if there are any issues, you need the plain data.

Also please rubocop.