Reports: Always clean up tmp directory when done
This restores old behavior of removing the tmp directory, even when the
report job fails. This was temporarily removed a while ago in commit
dc1b30ff to better understand problems with zipping up these tmp files.

To stop cluttering up the global tmp dir, we need to revert this change.