Create and use runtime directory for report jobs
systemd creates this directory for us (RuntimeDirectory=). Usually, it
would also clean it up when a service is stopped / restarted. We don't
want that (RuntimeDirectoryPreserve=yes), because of our long-running
zombie processes.

We make this switch because our use-case (continuously flushing huge CSV
files while generating them) more closely fits the purpose of these
runtime directories, as opposed to the system tmp dir.

Last but not least, we can now go back to a private tmp directory, as
our CSV files will now be stored in the (preserved) runtime directory.