Clean up leftover tmp files from killed reports
systemd can automatically clean up old files for us.

The options explained in detail:
- "e" is the mode: it marks the given directory for time-based cleanup
  (as the only other parameter that's given is the "age")
- the path is the runtime directory configured in the previous commit
- mode, owner and group are skipped - systemd already sets these
  correctly for any RuntimeDirectory
- an "age" of one day means that files older than one day will
  automatically be removed within the given directory, whenever the
  systemd-tmpfiles process does its thing

If you now say "Hey, reports can run for more than one day", you're
right. Thankfully, systemd-tmpfiles has an exclusion mechanism based on
"shared locks", which we implemented a while ago. As long as one of our
reports is running, it keeps a lock on its runtime directory, so that
the files will not be deleted by systemd-tmpfiles.

(#XO-1044)