Correct expansion for service environment args
The arguments passed to the `ExecStart` command of systemd services
were incorrectly parsed, so that an empty `*_ARGS` variable was not
stripped away but instead treated as a positional argument holding the
empty string.

This surfaced due to a fix for our internal wrapper script (which
sets up environment variables and working directory for each service)
to quote its own arguments. For example, without proper quoting, the
wrapper script runs `exec bundle exec rails puts "echo"` whereas,
with proper quoting, the arguments are correctly passed:
`exec bundle exec rails 'puts "echo"'`. See 10688c3a for the change.

For the `*-unicorn.service` command, this change means that unicorn
receives a positional argument (the empty string) and tries using it as
the path of the `config.ru` file to boot the application. The ruby code
to default to `config.ru` looks like this: `ARGV[0] || "config.ru"`. It
therefore only defaults to `"config.ru"` if *no* argument is given, but
not an empty string (which is `true`).

With the change, the arguments (e.g., `UNICORN_ARGS`) are split first
and then provided as separate parameters to the unicorn command.
See this systemd documentation for reference:
https://man.archlinux.org/man/systemd.service.5#COMMAND_LINES

Part of XI-5065