Increase enrollments per page limit
This CL increases the numbr of enrollments fetch per page in course
reports to 1000. As of now, the course service limits to 500, but the
report code appears to be able to handle larger pages.

Compared to the previous 50 enrollments per page, this cuts the total
number of requests by ten.