Fix linear progress for course reports
This CL changes the progress calculation for course reports to show a
more accurate linear progress.

A new utility class `Progress` manages a multi-value progress, where
multiple progresses can be managed at once, with their own current and
maximum values. This eases the implementation within the report as each
course only has to track its own progress. The class calculates a single
global progress, based on the individual progresses.

The course report now fetches a single enrollments page for each course
first, to determine the approximate total number of enrollments for each
course, to set the maximum for the progress. The numbers, including the
maximum, are updated after each processed enrollment. This might change
the total number, for example, when a new user enrolls into the course
while the report is running, but the overall progress will be pretty
accurate.

The course report is further split up into an `#each_run` and an
`#each_course` method. The first one will instantiation and course
iteration, the second one will yield the CSV lines for each course. This
change can further be extended to extract more common parts in base
classes or modules.