Rubocop: Don't disable MethodObjectAsBlock
The fix is easier than what Tobi seemed to assume in MR !275. Or maybe
this is what he meant by ugly. ;)

I prefer this to disabling another cop, though. And we might see a
refactoring of this main loop of each report anyway. At least Jan had a
proposal somewhere.