Add support for Redis Sentinel in Sidekiq
This lets Sidekiq work with a Redis Sentinel cluster, if desired, on
production systems.

The existing `config/sidekiq_redis.yml` file can now have a `sentinels`
key, with a list of sentinels to connect to:

    sentinels:
      - host: 'redis-a'
        port: 26381
      - host: 'redis-b'
        port: 26381
      - host: 'redis-c'
        port: 26381

Due to the nature of Sidekiq, write capabilities are needed on both the
client side (for publishing jobs) and on the server side (for marking
them as processed), so the "master" role is hardcoded.

(#XI-5437)