Fix date_filter for metrics
The `date_filter` option did not require the end date to be always set
if a start date is provided. The metrics using the `date_filter` so far
always provided the current date as a parameter. The `certificates`
metric did not.

This is now fixed in a generic way in the backend, not relying on the
parameters passed on from the API.

Fixes LANALYTICS-EV.
Regression from xikolo/lanalytics!545.