Sidekiq cron: Remove existing cron job definitions on initialize
Cron jobs defined via the `cron.yml` are loaded on initialization and
stored in Redis. Already existing cron job definitions are not cleaned
up. This can easily be forgotten when removing cron definitions (from
the code base) and leads to issues, trying to schedule jobs for
non-existing job classes.

To resolve this, we now clean up all definitions and reload it on
application restart.

This recently happened for the `CheckAllRulesWorker`:
See 6493edcf and https://openhpi.sentry.io/issues/5006983249.

Example:

```
> Sidekiq::Cron::Job.all
=>
[#<Sidekiq::Cron::Job:0x00007f3008e89600
  @active_job=false,
  @active_job_queue_name_delimiter="",
  @active_job_queue_name_prefix="",
  @args=[],
  @cron="*/15 * * * *",
  @date_as_argument=false,
  @description="",
  @fetch_missing_args=false,
  @klass="CheckAllRulesWorker",
  @last_enqueue_time=2024-04-16 11:45:04 UTC,
  @message="{\"retry\":false,\"queue\":\"high\",\"class\":\"CheckAllRulesWorker\",\"args\":[]}",
  @name="qc_rules_sync",
  @queue="high",
  @queue_name_with_prefix="high",
  @source="dynamic",
  @status="enabled",
  @symbolize_args=false>
]
```

See https://github.com/sidekiq-cron/sidekiq-cron for reference.

Part of XI-5427