feat: Use IterableJob for calculating course stats
Starting with v7.3, Sidekiq allows jobs to declare themselves as
iterable by using `Sidekiq::IterableJob`. This means that the job can
be decomposed into a sequence of elements to process; Sidekiq can stop
and restart the job anywhere within this sequence using a cursor.

This is especially useful for long-running jobs, allowing them to be
continued instead of retried completely.
See https://github.com/sidekiq/sidekiq/wiki/Iteration for details.

The course stat calculation is a typical long-running job, since stats
are calculated for each job, before sending out the daily statistic
email. This has been a huge pain point in the past.

The stats calculation is not efficient and involves many requests to
the course service for each and every course. The course services does
NOT persist statistic information, it at least caches some of it. The
information is persisted in the lanalytics service, but it still
issues requests to xi-course. Reducing the number of requests with
iterable jobs is a step to improve the situation.

Disclaimer: Each execution is expected to be completed within 25
seconds, which is the Sidekiq default for terminating jobs. This may
not be possible for course statistic calculations, but the operation
is idempotent. Only after successful responses from xi-course, the
statistics are updated. If not, i.e. in case of job termination, the
results are discarded and the calculation will be started again at a
later point.

This is not a full solution for the problem of slow course statistics
calculation. We still need more aggregated and persisted results, e.g.
by finishing the PLE implementation.

Relates to XI-6185