fix: Failing to load lanalytics custom classes
Rails 7.2 exclusively uses the new Zeitwerk loader, that does not accept
automatically loading classes, that originate from reloadable locations,
since reloading will fail with spurious results when they are used e.g.
from initializers. This broke the initializers that loaded the data
pipelines.

This commit wraps the initializer in a `#to_prepare` hook, which will
reload the pipeline set up whenever the relevant code is reloaded. This
at least fixes the immediate errors, but might still have some
unexpected changes in development mode, when code actually is reloaded
since it will call the set up again.

But this will only affect development for now, and might even work since
sources and pipelines are added to name-based hash, so when called
again, new objects will be placed at the same position in the hash.