fix(lint): Resolve Rails/ReversibleMigration offenses
Not all migrations are fully reversible, especially not with Rails 4.2.

For example, in Rails 4.2, using `-> { "CURRENT_TIMESTAMP" }` in
`change_column_default` is not supported. This syntax was introduced in
Rails 5.0.

In addition, there are data loss risks when rolling back some
migrations:
If any `annotation` values exceed 255 characters (the default limit for
string in Rails), they will be truncated when rolling back to string.

Still these are good examples how reversible migrations can be achieved.